+++ packages/provider-utils/src/index.ts
@@ -36,21 +36,10 @@ export {
   type InferSchema,
   type LazySchema,
   type Schema,
+  type ValidationResult,
 } from './schema';
 export * from './uint8-utils';
 export * from './validate-types';
-export {
-  asValidator,
-  isValidator,
-  lazyValidator,
-  standardSchemaValidator,
-  validator,
-  type FlexibleValidator,
-  type InferValidator,
-  type LazyValidator,
-  type ValidationResult,
-  type Validator,
-} from './validator';
 export { VERSION } from './version';
 export { withUserAgentSuffix } from './with-user-agent-suffix';
 export * from './without-trailing-slash';