+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/anthropic
 
+## 3.0.0-beta.21
+
+### Patch Changes
+
+- Updated dependencies [6306603]
+  - @ai-sdk/provider-utils@4.0.0-beta.16
+
+## 3.0.0-beta.20
+
+### Patch Changes
+
+- Updated dependencies [f0b2157]
+  - @ai-sdk/provider-utils@4.0.0-beta.15
+
 ## 3.0.0-beta.19
 
 ### Patch Changes