+++ packages/llamaindex/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/llamaindex
 
+## 2.0.0-beta.42
+
+### Patch Changes
+
+- ai@6.0.0-beta.42
+
+## 2.0.0-beta.41
+
+### Patch Changes
+
+- ai@6.0.0-beta.41
+
 ## 2.0.0-beta.40
 
 ### Patch Changes