+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/openai
 
+## 3.0.0-beta.25
+
+### Patch Changes
+
+- Updated dependencies [6306603]
+  - @ai-sdk/provider-utils@4.0.0-beta.16
+
+## 3.0.0-beta.24
+
+### Patch Changes
+
+- Updated dependencies [f0b2157]
+  - @ai-sdk/provider-utils@4.0.0-beta.15
+
 ## 3.0.0-beta.23
 
 ### Patch Changes